restore sidebar and move meta links in a dropdown
this is the result of an ad-hoc usability observation I did at StudioXX:
the first thing the user did when browsing the site was to hover their
mouse over to the top right links. they didn't click anything because
they didn't understand what the links were.

this is prime screen real-estate. it should be contact information, the
blog and maybe *one* other link, that's all.

we move the edit, comments and all that crap in a "meta" link down
below.

note the BOOTMENU (which I have no idea wtf it is) is completely removed
from the list, because it is already in the header.