revert back to using CDNs for bootstrap and jquery
I understand there are security and privacy implications with this,
but I consider performance considerations trump those. Furthermore,
the issues are limited.

Security-wise, the content is exactly the same as it was before, and
integrity is checked by the client, thanks to the `integrity`
attribute. I have taken checksums from the versions in this git
repository and they match files downloaded from the CDN at the time of
writing.

Privacy-wise, I have asked the MaxCDN people (behind the Jquery and
bootstrap CDNs) about their privacy policies. They seem acceptable:

 * they keep 7 days of logs
 * they do not keep user-identifiable information (e.g. cookies)
 * they do not resell data to third-parties

Their privacy policy is here: https://www.maxcdn.com/legal/

While it *does* mean they keep logs, I also do on my own server, so I
think this is an acceptable compromise to give the site a small
performance boost.

The actual static files are kept locally because they may be
referenced in HTML files that could still be in caches. They will be
removed once I am satisfied with the performance and results.