security update of jQuery to 1.12.4
according to [Snyk], 1.11.2 is vulnerable to multiple XSS attacks. It
is unclear if 1.12.4 solves all issues, particularly
npm:jquery:20150627, which requires 3.0 to fix. Unfortunately that
would mean a Bootstrap upgrade as well, so it's actually harder to
pull.

If I'm going to upgrade bootstrap, I'm probably going to redo the
design in something else that's simpler than all this junk.

 [Snyk]: https://snyk.io/vuln/npm:jquery