expand favicon support to non-standard devices
Turns out that the good old favicon.ico is not enough anymore. Apple,
Microsoft and Google all thought they had *better* ideas so they
designed non-standard extensions to the "favicon" standard.

Adding compatibility for this requires more or less 7 images and two
metadata file, only 4 of which are actually linked from the HTML - the
rest being "automagically" fetched depending on the platform/device.

This is a shame and an embarrassment for the open web, but I prefer to
have pretty icons for users than try to fight such a useless battle.

This obviously requires changes to websites using the theme, but should
still work without those extra configs: those will just yield a 404. The
previous favicon.png, however, will fail to be found anymore in most
cases, so that image should at least be converted to a favicon.ico or
similar.

Cargo-culted from https://realfavicongenerator.net/