figures formatting
We use figures because it features the `figcaption` element that allows
us to add a caption without reverting to messy table or div styling. A
todo item was opened upstream to see how this could be improved in
ikiwiki's img directive itself, but since we don't generally use it, we
don't need to hack that directly and instead inject the elements
directly in the Markdown, and they get through the scrubber without
problems.

The align-right style is for images with want to float to the right of
text. This required display: table otherwise the float does not work and
the elements just stay where they are if there's a caption. The padding
is necessary otherwise everything is all crammed up, and so is the
max-width: this might resize some images, but if that's a problem, they
should be kept centered as they are just too big.

Thanks to joe on IRC for the help with that latter bit, that was epic.