add breadcrumbs for non top-level pages
This is essential for pages like "services" or "hardware" otherwise you
can't browse back to the upper pages without using the "back" button. If
you were hotlinked in, there is no way back at all. This will also be
useful to link back to the blog, although there is already a link there.
Might be worth making an exception for entries in the top-level links.

This looks *kind* of okay, but I might get tired of it again...